Replace gomail/gomail with wneessen/go-mail in email mail utility (#4867)
The gomail/gomail library was replaced with the go-mail library for sending emails. This change involved updating the go.mod, go.sum, and email.go files. The email.go changes include a new method of setting TO and FROM addresses, and a new way to dial and send the email. This switch to go-mail provides better error handling and more customization options.

Co-authored-by: Jiahui <4543bxy@gmail.com>